requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
dnspython==2.6.1
frozenlist==1.4.1
idna==3.6
motor==3.3.2
multidict==6.0.5
pydantic==2.6.3
pydantic_core==2.16.3
pymongo==4.6.2
pyTelegramBotAPI==4.16.1
python-dateutil==2.9.0
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
typing_extensions==4.10.0
urllib3==2.2.1
yarl==1.9.4
